+++ src/components/ActiveCombatant.vue
@@ -1,7 +1,7 @@
 <template>
     <div id="active-combatant">
         <h2 class="title">Active Combatant</h2>
-        <combatant-view v-if="activeCombatant && combatStarted" :combatant="activeCombatant"></combatant-view>
+        <combatant-view :component-max-width="370" v-if="activeCombatant && combatStarted" :combatant="activeCombatant"></combatant-view>
         <div v-else>
             Click Start Encounter to determine the round order. The stat-block for the Active Combatant will be displayed here.
         </div>