+++ src/store.js
@@ -2,12 +2,14 @@ import Vuex from "vuex";
 import TraitsAndTalents from "@/classes/TraitsAndTalents";
 import {Combatant, NPC} from "@/classes/Combatant";
 import Vue from 'vue';
+import Roller from "@/classes/Roller";
 
 Vue.use(Vuex);
 export const DEFAULT_SETTINGS = {
     initiativeType: 'default',
     useMaxAdvantage: false,
     autoAdvantageOnWounds: true,
+    grouped: true,
 }
 
 const normalisationPerStateVersion = {
@@ -49,15 +51,18 @@ export const store = new Vuex.Store({
             characters: {},
             encounters: {},
         },
+
+        // Combat
         combatants: [],
+        activeCombatant: null,
+        combatStarted: false,
+        combatRound: 1,
 
         // Selections
-        activeCombatant: null,
         selectedCombatant: null,
 
         // View state
         loading: true,
-        combatStarted: false,
         libraryOpen: false,
         libraryTab: 0,
 
@@ -65,7 +70,19 @@ export const store = new Vuex.Store({
         ...DEFAULT_SETTINGS
     },
     getters: {
-        allTraits() {return TraitsAndTalents.TRAITS},
+        allTraits(state) {return TraitsAndTalents.TRAITS},
+        combatantsWithNumbers(state) {
+            let idsFound = [];
+            let ids = [];
+            state.combatants.forEach(combatant => {
+                if (idsFound.includes(combatant.id) && !ids.includes(combatant.id)) {
+                    ids.push(combatant.id);
+                } else {
+                    idsFound.push(combatant.id);
+                }
+            })
+            return ids;
+        },
     },
     mutations: {
         openLibrary(state) {
@@ -84,7 +101,23 @@ export const store = new Vuex.Store({
             state.selectedCombatant = null;
         },
         saveCombatant(state, combatant) {
-            Vue.set(state.library[combatant instanceof NPC ? 'bestiary' : 'characters'], combatant.id, combatant);
+            let type = combatant instanceof NPC ? 'bestiary' : 'characters';
+            Vue.set(state.library[type], combatant.id, combatant);
+            if (combatant.isUnique !== false) {
+                let combatantInState = state.library[type][combatant.id];
+                let index = state.combatants.findIndex(icc => icc.id === combatantInState.id);
+                if (index > -1) {
+                    combatantInState.initiative = state.combatants[index].initiative;
+                    Vue.set(state.combatants, index, combatantInState);
+
+                    if (state.selectedCombatant.id === combatantInState.id) {
+                        state.selectedCombatant = state.combatants[index];
+                    }
+                    if (state.activeCombatant.id === combatantInState.id) {
+                        state.activeCombatant = state.combatants[index];
+                    }
+                }
+            }
         },
         destroyCombatant(state, combatant) {
             Vue.delete(state.library[combatant instanceof NPC ? 'bestiary' : 'characters'], combatant.id);
@@ -138,6 +171,20 @@ export const store = new Vuex.Store({
         setCombatantAdvantage(state, {combatant, advantage}) {
             let combatantInState = state.combatants.find(com => com === combatant);
             Vue.set(combatantInState, 'advantage', advantage);
+        },
+        toggleCombat(state) {
+            state.combatStarted = !state.combatStarted;
+            if (state.combatStarted) {
+                state.combatRound = 1;
+                state.activeCombatant = state.combatants[0];
+            }
+        },
+        setCombatants(state, combatants) {
+            state.combatants = combatants;
+        },
+        setCombatantInitiative(state, {combatant, initiative}) {
+            let combatantInState = state.combatants.find(com => com === combatant);
+            Vue.set(combatantInState, 'initiative',  initiative);
         }
     },
     actions: {
@@ -206,6 +253,95 @@ export const store = new Vuex.Store({
         advantageToZero(context, combatant) {
             context.commit('setCombatantAdvantage', {combatant: combatant, advantage: 0});
         },
+        toggleCombat(context) {
+            let combat_starts = !context.state.combatStarted;
+            if (combat_starts) {
+                store.dispatch('determineInitiative').then(() => context.commit('toggleCombat'));
+            }
+            else {
+                context.state.combatants.forEach(combatant => {
+                    context.commit('setCombatantAdvantage', {combatant: combatant, advantage: 0});
+                })
+                context.commit('toggleCombat');
+            }
+        },
+        determineInitiative(context) {
+            let combatantsByInitiative = [];
+
+            let initiativeType = context.state.initiativeType;
+            // Step 1. Determine the initiative of each combatant
+            context.state.combatants.forEach(combatant => {
+                let initiative = combatant.getInitiative(initiativeType);
+                if (initiativeType === 'test') {
+                    if (initiative.success_levels !== 0 || initiative.success) {
+                        initiative = initiative.success_levels;
+                    } else {
+                        initiative = -0.001
+                    }
+                }
+                context.commit('setCombatantInitiative', {combatant: combatant, initiative: initiative});
+                if (!combatantsByInitiative[initiative]) {
+                    combatantsByInitiative[initiative] = [];
+                }
+                combatantsByInitiative[initiative].push(combatant);
+            })
+
+            // Step 2. Resolve ties
+            let orderedCombatants = [];
+            Object.keys(combatantsByInitiative).sort((a, b) => b - a).forEach(initiative => {
+                if (combatantsByInitiative[initiative].length > 1) {
+                    let tiedCombatants = combatantsByInitiative[initiative];
+                    let tiedCombatantsById = {};
+
+                    let tiesToSolve = [];
+                    if (context.state.grouped) {
+                        tiedCombatants.forEach(combatant => {
+                            if (!tiedCombatantsById[combatant.id]) {
+                                tiedCombatantsById[combatant.id] = [];
+                            }
+                            tiedCombatantsById[combatant.id].push(combatant);
+                        })
+
+                        let ids = Object.keys(tiedCombatantsById);
+                        if (ids.length > 1) {
+                            ids.forEach(id => {
+                                tiesToSolve.push(tiedCombatantsById[id][0]);
+                            })
+                        }
+                        else {
+                            let combatantsToAdd = Object.values(tiedCombatantsById)[0];
+                            if (combatantsToAdd.length > 0) {
+                                combatantsToAdd.sort((a, b) => a.no - b.no);
+                            }
+                            combatantsToAdd.forEach(combatant => {
+                                orderedCombatants.push(combatant);
+                            })
+                        }
+                    } else {
+                        tiesToSolve = tiedCombatants;
+                    }
+
+                    tiesToSolve
+                        .sort((a, b) => {
+                            let results = Roller.opposedTest(a.stats.agi.value, b.stats.agi.value);
+                            return results.winner === 'A' ? -1 : 1;
+                        })
+                        .forEach(combatant => {
+                            if (context.state.grouped) {
+                                tiedCombatantsById[combatant.id].sort((a, b) =>a.no ? a.no - b.no : 0).forEach(combatant => orderedCombatants.push(combatant));
+                            } else {
+                                orderedCombatants.push(combatant);
+                            }
+                        })
+                } else {
+                    orderedCombatants.push(combatantsByInitiative[initiative][0])
+                }
+            })
+
+            console.log(orderedCombatants);
+            context.commit('setCombatants', orderedCombatants);
+        },
+
     },
 });