+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add filter for unique/non-unique NPCs in bestiary
 - Add state versioning and backwards compatibility
 - Set `updated_at` whenever it is saved
+- Updating a unique combatant updates their data throughout the application
 
 ##[0.4.0]
 - Further implement Vuex, major refactor of use of variables and saving of state